requirements.txt
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.3.0
Pillow>=8.0.0
einops>=0.3.0
torchsummary>=1.5.0 